+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Text blocks content is now output using the `extra` markdown flavor.
 * The sprig plugin now hosts the hmtx library locally, so removed the `htmxScriptUrl` custom setting.
 * Added the `prepareText` twig filter and corresponding custom settings.
+* Add a preview target by default if on Craft Pro.
 
 ## 3.1.0 2023-04-06